cleanup tests utils
