+++ x/sequencers/testutils/testutils.go
@@ -2,35 +2,21 @@ package testutils
 
 import (
 	"testing"
-	"time"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
-	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	"github.com/cosmos/cosmos-sdk/store"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	typesparams "github.com/cosmos/cosmos-sdk/x/params/types"
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	"github.com/stretchr/testify/require"
 	"github.com/tendermint/tendermint/libs/log"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 	tmdb "github.com/tendermint/tm-db"
 
-	"github.com/dymensionxyz/rollapp/app"
 	"github.com/dymensionxyz/rollapp/x/sequencers/keeper"
 	"github.com/dymensionxyz/rollapp/x/sequencers/types"
 )
-
-func NewTestContext() sdk.Context {
-	return sdk.NewContext(nil, tmproto.Header{}, false, log.NewNopLogger())
-}
-
-func NewTestSequencerKeeperFromApp(t *testing.T, app *app.App) (*keeper.Keeper, sdk.Context) {
-	k := &app.SequencersKeeper
-	ctx := app.BaseApp.NewContext(false, tmproto.Header{Height: 1, ChainID: "rollapp-1", Time: time.Now().UTC()})
-	return k, ctx
-}
 
 func NewTestSequencerKeeper(t *testing.T) (*keeper.Keeper, sdk.Context) {
 	storeKey := sdk.NewKVStoreKey(types.StoreKey)
@@ -57,16 +43,9 @@ func NewTestSequencerKeeper(t *testing.T) (*keeper.Keeper, sdk.Context) {
 		paramsSubspace,
 	)
 
-	ctx := NewTestContext().WithMultiStore(stateStore)
+	ctx := sdk.NewContext(nil, tmproto.Header{}, false, log.NewNopLogger()).WithMultiStore(stateStore)
 	// Initialize default params
 	k.SetParams(ctx, types.DefaultParams())
 
 	return k, ctx
-}
-
-// NewValidator is a testing helper method to create validators in tests
-func NewValidator(t testing.TB, operator sdk.ValAddress, pubKey cryptotypes.PubKey) stakingtypes.Validator {
-	v, err := stakingtypes.NewValidator(operator, pubKey, stakingtypes.Description{})
-	require.NoError(t, err)
-	return v
 }
